requirements.txt
PyMuPDF
datasets
nltk
unsloth

marker-pdf[full]
jsonlines 
tqdm